Re-organizing bundles

+++ census.gov/census.gov-2000_population-sf1geo-orig/bundle.py
@@ -17,6 +17,8 @@ class Bundle(UsCensusBundle):
     def prepare(self):
         '''Create the prototype database'''
   
+        #self.clean()
+  
         if not self.database.exists():
             self.database.create()
 
@@ -89,16 +91,17 @@ class Bundle(UsCensusBundle):
         state. Then creates a partition for each of the geo files. '''
 
         import yaml 
-      
-        urls = yaml.load(file(self.urls_file, 'r')) 
           
         # Process the geo files. 
+        urls = yaml.load(file(self.urls_file, 'r')) 
         for state, source in urls['geos'].items():
             self.load_geo(state, source)
     
         self.combine_geo()
     
-        self.split_geo_sqlite()
+        #self.split_geo_sqlite()
+    
+        self.split_geo()
     
         return True
     
@@ -114,33 +117,35 @@ class Bundle(UsCensusBundle):
         pid = PartitionIdentity(self.identity, space=state)
         partition = self.partitions.find(pid)
        
-        if self.library.get(partition.identity.id_):
+        if self.library.get(partition.identity):
             self.log("Found in library, skipping. "+pid.name)
             return
    
         if not partition.database.exists():
+            self.log("Create database "+partition.database.path)
             partition.create_with_tables('sf1geo')
+        else:
+            self.log("Database already exists "+partition.database.path)
 
-        def print_follow(row):
-            print row
-            
+     
         processors = [CensusTransform(c, useIndex = True) for c in table.columns ]
             
         def process_row(row):
             return [ f(row) for f in processors ]
-
+  
         retry = 4
         while retry > 0:
             retry -= 1
             #try:  
             with self.filesystem.download(source) as zip_file:
                 with self.filesystem.unzip(zip_file) as rf:
+    
                     self.log("Processing GEO file: "+rf)
                               
                     db_path = partition.database.path
-                    self.log("  Partion: "+str(partition))
+                    self.log("  Partion: "+partition.name)
                     self.log("  To Database: "+db_path)
-                     
+                               
                     t = petl.fromregex(rf, regex=regex, header=header) #@UndefinedVariable
  
                     if state == 'pr': # Some unicode characters, but only in this file
@@ -148,13 +153,17 @@ class Bundle(UsCensusBundle):
 
                     (t
                      .mogrify(lambda row: process_row(row))
-                     .progress(100000)
+                     .progress(10000)
                      .appendsqlite3(db_path,'sf1geo'))
-                          
+                              
                     dest = self.library.put(partition)
                     self.log("Install in library: "+dest)
+                    
+                    if not self.library.get(partition.identity):
+                        raise Exception("Library Installation failed: "+partition.identity)
+                    
                     partition.database.delete()
-                                  
+            
                     return True
                 
         return False
@@ -168,8 +177,8 @@ class Bundle(UsCensusBundle):
 
         combined = self.partitions.find(PartitionIdentity(self.identity, table='sf1geo'))
      
-        if not combined.database.exists():
-            combined.create_with_tables(combined.table.name)
+        combined.database.delete()
+        combined.create_with_tables(combined.table.name)
 
         if self.library.get(combined.identity.id_):
             self.log("Found in library, skipping. "+combined.identity.name)
@@ -201,17 +210,13 @@ class Bundle(UsCensusBundle):
             if not geo.identity.space:
                 continue; # Only take state file; ignore national split files
             
-
             name = cdb.attach(geo.database)
             
             sf1t = geo.schema.table('sf1geo')
-           
-            map_ = { "TRIM({})".format(c.name): c.name for  c in sf1t.columns}
-            
+        
             cdb.copy_from_attached(('sf1geo','sf1geo'), 
                                    name=name,
-                                   columns=map_,
-                                on_conflict = 'ABORT')
+                                   on_conflict = 'ABORT')
             cdb.detach(name)
  
 
@@ -219,9 +224,7 @@ class Bundle(UsCensusBundle):
         self.log("Installed combined file in library: "+dest)
 
     def split_geo_sqlite(self):
-        '''Split the geo file using attachment and table copy'''  
-    
-                
+        '''Split the geo file using attachment and table copy'''         
         from databundles.partition import PartitionIdentity
 
         #
@@ -237,6 +240,20 @@ class Bundle(UsCensusBundle):
 
         i = 0
         
+        skips = set()
+        
+        for table in self.schema.tables:
+            p = self.partitions.find(
+                        PartitionIdentity(self.identity, table=table.id_))
+        
+            if self.library.get(p):
+                skips.add(table)
+            
+        if len(skips) == len(self.schema.tables):
+            self.log("All Split tables alreay in library, skipping" )
+            return True
+
+    
         for result in q.all:   
 
             i += 1
@@ -254,7 +271,8 @@ class Bundle(UsCensusBundle):
 
             for table in self.schema.tables:
                 
-                if table.name in ['sf1geo']:
+              
+                if table.name in ['sf1geo'] or table in skips:
                     continue
 
                 p = self.partitions.find(
@@ -295,8 +313,140 @@ class Bundle(UsCensusBundle):
             dest = self.library.put(p)
             self.log("Installed in library: "+dest)
             p.database.delete()
+  
+    def split_geo(self):
+        '''Split the geo file into seperate tables'''
+        from databundles.database import  insert_or_ignore
+        from databundles.partition import PartitionIdentity
+        from databundles.transform import PassthroughTransform
+        import time
+        import hashlib
+        from collections import OrderedDict 
+ 
+        pid = PartitionIdentity(self.identity, table='sf1geo')
+        p = self.partitions.find(pid) # Identity needs to have id_
+        combined = self.library.get(p.identity)
+   
+        if not combined:
+            raise Exception("Didn't get Combined bundle from library "+pid.path)
+   
+        tables = []
+        
+        # Must have record_code first so
+        for table in self.schema.tables:
+            if table.name == 'record_code':
+                tables.insert(0, table)
+            elif table.data.get('split_table', False):
+                tables.append(table)
+
+      
+        for table in tables:
+           
+            row_i = 0
+            t_start = time.clock()
+            value_set = OrderedDict()
+            id_set = {}
+
+            partition = self.partitions.find(
+                    PartitionIdentity(self.identity, table=table.id_))
             
-    
+            if not partition.database.exists():
+                partition.create_with_tables(table.name)
+         
+            # Get all of the source columns, but exclude the foriegn keys
+            source_cols = ([c.name for c in table.columns 
+                            if not ( c.name.endswith('_id') and not c.is_primary_key)
+                           ])
+                
+            processors = [PassthroughTransform(c) for c in table.columns if c.name in source_cols ] 
+            processors[0] = lambda row: 95959 # replace the primary key with a placeholder
+                 
+            columns = [c for c in table.columns if c.name in source_cols ]
+                
+            self.log(" ")
+            self.log(" ---- " + table.name)
+            for row in combined.database.connection.execute("SELECT * FROM sf1geo"):
+            
+                row_i += 1
+                
+                if row_i % 10000 == 0:
+                    self.ptick('.')
+               
+                if row_i % 100000 == 0:
+                    self.ptick(str(row_i/100000))
+                    self.ptick("E5 ")
+                    self.ptick(str(int( row_i/(time.clock()-t_start))))
+                    self.ptick(' ')
+                    self.ptick(str(len(value_set)))   
+
+                if row_i % 500000 == 0:
+                    break
+
+                values=[ f(row) for f in processors ]
+               
+                m = hashlib.md5()
+                for x in values:
+                    m.update(str(x))
+         
+                hash = m.digest()
+         
+                if hash not in id_set:
+                    value_set[hash] = values
+                    id_set[hash] = [(row['stusab'], row['logrecno'])]
+                else:
+                    id_set[hash].append((row['stusab'],row['logrecno']))
+           
+            
+            self.log(' ')
+            self.log("Length: "+str(len(value_set)))  
+            self.log('Collection Write {}/s '.format(int( len(value_set)/(time.clock()-t_start))))
+
+
+            # This generator expressions replaces the first value in the row, the placeholder
+            # primary key, with an acualt key. 
+          
+            ins_gen = ([i]+item[1][1:] for i, item in enumerate(value_set.iteritems()))
+
+            ins = insert_or_ignore(table.name, columns)
+          
+            t_start = time.clock()
+            db = partition.database
+            self.log('Write to db: '+db.path)
+            t_start = time.clock()
+            db.dbapi_cursor.executemany(ins, ins_gen)
+            db.dbapi_connection.commit()
+            db.dbapi_close()
+            self.log('Db Write {}/s '.format(int( len(value_set)/(time.clock()-t_start))))
+        
+            if table.name != 'record_code':
+                #
+                # Update the record_codes table to make the logrecno and stusab
+                # keys to the new table rows. 
+                
+                def update_gen(): 
+                    for i, item in enumerate(value_set.iteritems()):
+                        for id in id_set[item[0]]:
+                            yield (i,id[0], id[1])
+
+                rcp = self.partitions.find(
+                    PartitionIdentity(self.identity, table='record_code'))
+            
+                update = ("update record_code set {}=? where stusab = ? and logrecno = ?"
+                           .format(table.name+'_id' ) )
+                
+                print update
+                g = update_gen()
+            
+                print rcp.database.path
+                db = rcp.database
+                t_start = time.clock()
+                db.dbapi_cursor.executemany(update, g)
+                db.dbapi_connection.commit()
+                db.dbapi_close()
+                self.log('Record Code Write {}/s '.format(int( len(value_set)/(time.clock()-t_start))))
+
+        return True
+
 import sys
 
 if __name__ == '__main__':
